+++ client/src/interceptors/index.js
@@ -1,14 +1,14 @@
-import { HTTP } from "../util";
+import toastr from "toastr";
 
-/**
- *  We can add request/response interceptors here. I.e. to 
- *  replace the locally stored auth token.
- */
-
-HTTP().interceptors.response.use(response => {
-    console.log("[interceptor] Intercepted the response!");
+export function success( response )
+{
+    console.log( "[interceptor] Success: ", response );
     return response;
-}, error => {
-    console.log("[intercetpor] Intercepted an error response.");
-    return Promise.reject(error);
-});
+}
+
+export function error( response )
+{
+    toastr.error(`Info: ${response.message}`, "Fehler");
+    console.error( "[interceptor] Error: ", response );
+    return response;
+}