+++ server/models/Product.js
@@ -9,6 +9,7 @@ var productSchema = mongoose.Schema({
     weight: { type: Number, required: true, default: 500 }, // in gramm
     price: { type: Number, required: true },
     public: { type: Boolean, required: true, default: false },
+    isInPreparation: { type: Boolean, required: true, default: true },
     hasPorto: { type: Boolean, required: true, default: false },
     image: { type: String, required: false, default: "" },
     sortIndex: { type: Number, required: true, default: 0 },